requirements.txt
# Core dependencies
requests>=2.25.0
PyYAML>=5.4.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.3.0
flake8>=4.0.0
mypy>=0.950
types-requests>=2.25.0
types-PyYAML>=5.4.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
